decoupling Utility/common into file and visualization
