+++ examples/Python/Utility/file.py
@@ -2,9 +2,7 @@
 # The MIT License (MIT)
 # See license file or visit www.open3d.org for details
 
-import copy
 import re
-from open3d import *
 import os
 from os import listdir, makedirs
 from os.path import exists, isfile, join, splitext
@@ -70,37 +68,3 @@ def get_rgbd_file_lists(path_dataset):
 def make_folder(path_folder):
     if not exists(path_folder):
         makedirs(path_folder)
-
-
-#######################
-# visualization related
-#######################
-flip_transform = [[1, 0, 0, 0], [0, -1, 0, 0], [0, 0, -1, 0], [0, 0, 0, 1]]
-
-def draw_geometries_flip(pcds):
-    pcds_transform = []
-    for pcd in pcds:
-        pcd_temp = copy.deepcopy(pcd)
-        pcd_temp.transform(flip_transform)
-        pcds_transform.append(pcd_temp)
-    draw_geometries(pcds_transform)
-
-
-def draw_registration_result(source, target, transformation):
-    source_temp = copy.deepcopy(source)
-    target_temp = copy.deepcopy(target)
-    source_temp.paint_uniform_color([1, 0.706, 0])
-    target_temp.paint_uniform_color([0, 0.651, 0.929])
-    source_temp.transform(transformation)
-    source_temp.transform(flip_transform)
-    target_temp.transform(flip_transform)
-    draw_geometries([source_temp, target_temp])
-
-
-def draw_registration_result_original_color(source, target, transformation):
-    source_temp = copy.deepcopy(source)
-    target_temp = copy.deepcopy(target)
-    source_temp.transform(transformation)
-    source_temp.transform(flip_transform)
-    target_temp.transform(flip_transform)
-    draw_geometries([source_temp, target_temp])
