+++ examples/Python/Utility/visualization.py
@@ -1,0 +1,39 @@
+# Open3D: www.open3d.org
+# The MIT License (MIT)
+# See license file or visit www.open3d.org for details
+
+from open3d import *
+import copy
+
+#######################
+# visualization related
+#######################
+flip_transform = [[1, 0, 0, 0], [0, -1, 0, 0], [0, 0, -1, 0], [0, 0, 0, 1]]
+
+def draw_geometries_flip(pcds):
+    pcds_transform = []
+    for pcd in pcds:
+        pcd_temp = copy.deepcopy(pcd)
+        pcd_temp.transform(flip_transform)
+        pcds_transform.append(pcd_temp)
+    draw_geometries(pcds_transform)
+
+
+def draw_registration_result(source, target, transformation):
+    source_temp = copy.deepcopy(source)
+    target_temp = copy.deepcopy(target)
+    source_temp.paint_uniform_color([1, 0.706, 0])
+    target_temp.paint_uniform_color([0, 0.651, 0.929])
+    source_temp.transform(transformation)
+    source_temp.transform(flip_transform)
+    target_temp.transform(flip_transform)
+    draw_geometries([source_temp, target_temp])
+
+
+def draw_registration_result_original_color(source, target, transformation):
+    source_temp = copy.deepcopy(source)
+    target_temp = copy.deepcopy(target)
+    source_temp.transform(transformation)
+    source_temp.transform(flip_transform)
+    target_temp.transform(flip_transform)
+    draw_geometries([source_temp, target_temp])
